use a system.data.js name vs system-attrs.data.js.  the -attrs made sense for unit-attrs b/c those are distinct from a unit, not so for systems.
